+++ src/main/java/ru/yandex/practicum/filmorate/dal/interfaces/ReviewRepository.java
@@ -15,7 +15,7 @@ public interface ReviewRepository {
 
     Optional<Review> getById(Long id);
 
-    Collection<Review> getReviewsByFilmId(Long filmId, Long count);
+    Collection<Review> getReviewsByFilmId(Long filmId, Integer count);
 
     void addLike(Long id, Long userId);
